+++ package.js
@@ -8,8 +8,8 @@ Package.describe({
 })
 
 Package.onUse(function (api) {
-  api.versionsFrom('2.0')
-  api.use(['meteor', 'ecmascript', 'check', 'mongo', 'typescript'])
+  api.versionsFrom('2.3')
+  api.use(['meteor', 'ecmascript', 'check', 'mongo', 'typescript', 'callback-hook'])
   api.use(['aldeed:collection2@3.5.0', 'aldeed:schema-index@3.0.0', 'aldeed:schema-deny@3.0.0'])
   // 'socialize:base-model@1.1.2'